Oh no

+++ ConsoleRPG/City.cpp
@@ -13,7 +13,7 @@ City::City(char* pName)
 
 City::~City()
 {
-	for (size_t i = 0; i < cityDestinationVector.size(); i++)
+	for (int i = 0; i < cityDestinationVector.size(); i++)
 	{
 		delete cityDestinationVector.at(i);
 	}
